tools/kconfig: Add stm32_ioconf() and stm32 pin names.
Added stm32_ioconf() which is largely the same as pic32_ioconf()
except that pin names have the format of P[A-Z][0-9][0-9]* and
since an stm32 microcontroller can have up to eleven gpio ports
just make it a nice round 16 ports for potential future expansion.

Modify the lexer regular expression for a pin to [PR][A-Z][0-9][0-9]*
which will accept potential pin names from [PR]A0 to [PR]Z255.

Get rid of the needless "#if ARCH_{machine}" conditional directives.
They were added to config over 40 years ago and broken from the start,
as they always evaluated to true.