sys/stm32: Allow both active-high and active-low buttons.
button1_pressed() now also works with buttons that are active-low,
like on the Nucleo boards.
'BUTTON_USER_PRESSED' is defined for each board/button as the state
that the button is in while being pressed; 0 for Nucleo, 1 for the
STM32F469I-Disco. This is compared against BSP_PB_GetState().

'BUTTON_USER_PRESSED' may be replaced with the Config gpio pins
'invert' functionality in a future commit, but this is solid as is.