sys/stm32: Add LL_GPIO_SetAFPin() to LL_GPIO library.
Add the unofficial extension LL_GPIO_SetAFPin() to the LL_GPIO library
and use the implementation throughout the stm32 port.

This new library function chooses to use either LL_GPIO_SetAFPin_0_7()
or LL_GPIO_SetAFPin_8_15() when setting a pin, as the case may be,
so that drivers can use the same API interface when setting GPIO pins.