sys/stm32: Set minimums for kernel/user flash/ram.
A minimum STM32 system can easily get by with 128kb flash, 128kb ram
with all the flash and 32kb of ram for the kernel, and 96kb user ram.
Currently, the F469IDISCO kernel is a bit over 91kb of flash and
23.7kb of ram (with 6kb more for u. and u0., for a total 29.7kb ram).
96kb user ram easily fits all executables in the system, with uucico
being the largest at 89.1kb.

Selectively use the 64kb CCM data ram for kernel ram if it is available,
else just use the last 32kb of regular ram (SRAM2) for kernel ram.

Remove USER_FLASH_* as it is not used for the STM32 port.