sys/pic32: Only check kernel flash area in badkaddr().
badkaddr() had an error in that it checks all of flash to see if the
supplied memory address is within the 512kb of flash in the PIC32MX.
Instead, it should only check if the supplied address is within the
defined kernel flash area, which is the first 192kb of flash.

Added KERNEL_FLASH_END define that is the correct end of kernel flash
and used it in badkaddr() to check for kernel flash addresses.

The value of USER_FLASH_START is unchanged, as it is the same as
KERNEL_FLASH_END.