sys/{kern,stm32}: Use linker script memory address symbols.
Memory addresses for start and end locations of flash and RAM are
defined only in the MEMORY section in the kernel linker script, and
are used throughout the kernel.
The specific symbols now used are:
  __kernel_flash_start
  __kernel_flash_end
  __kernel_data_start
  __kernel_data_end
  __user_data_start
  __user_data_end

The new symbols are publicly available through the sys/systm.h header.

The STM32 port now uses these new symbols, but the PIC32 port is still
using the older #defines from pic32/machparam.h, so temporarily define
in sys/systm.h the new symbols as aliases to the old defines on PIC32.

All kernel code is updated to use the new symbols, and are type casted
to the required type, as the case may be, as these symbols do not have
an inherent type but are declared as a 'char __symbol_name[]' for ease
of use and to not require to take the address of the symbol when used.

The MEMORY section now has a required memory region for USERRAM, which
for the STM32 port is at memory address 0x20000000 and must be at least
96 kbytes long. This memory region is used exclusively by userland.

USER_DATA_* #defines, stacktop(), and stackbas() in stm32/machparam.h
are kept for now, as they are needed and hardcoded by ps, w, and smlrc.
All other memory address #defines in stm32/machparam.h have been removed.
CCM RAM usage is only determined by linker script MEMORY section regions.

The userland start and end addresses are no longer hardcoded in stm32
startup code, and instead use the new memory address symbols.

There is no functional or binary change in the code.