sys/stm32: Add SPI2 SD card config for F4DISCOVERY board.
Bringing back the F4DISCOVERY board, and the first step is to enable
the SD card on SPI2 by adding the defines to stm32f4xx_nucleo_sd.h.
CS on PB12, SCK on PB13, MISO on PB14, and MOSI on PB15.

Cleaned up some of the noise while there.