sys/stm32: Support STM32 NUCLEO-F411RE development board.
The NUCLEO-F411RE development board has a 100MHz STM32F411RET6 MCU,
128 Kbytes of SRAM, 512 Kbytes of Flash, a Virtual Com Port via the
USB port (directly connected to USART2), an Arduino Uno Rev 3 connector,
ST Morpho extension headers, and a User button and a User LED.

An SD card (required for a root filesystem) can be connected through
an external board (such as the Adafruit MicroSD Breakout Board+, ID 254)
on the SPI1 port available on the Arduino connector on pins D10-D13.
The kernel for this board is configured for this very situation.

Add f411renucleo subdirectory to sys/stm32 and populate with Config,
kconfig-generated Makefile, and kernel memory linker file.

Memory is split up as follows: 32 Kbytes for kernel RAM where 26 Kbytes
is for the kernel and 3 Kbytes for u. area and 3 Kbytes for u0. area.
96 Kbytes (minus 0x80) for User RAM to run executables. This is the
minimum for the system (all executables run in under 96 Kbytes of RAM).
The 'minus 0x80' is due to a current bug that happens when the exec
loader is loading a large executable, such as cron, during boot.

The f411renucleo Config has the console device through USART2 at 115200
baud on GPIO pins PA2 and PA3 on AF7, which is directly connected to
the USB VCP through the embedded ST-LINK/V2-1 on the development board.

The User button is configured through Config; the User LED is added to
Config but it is commented out, as it conflicts with SPI1 SCK on D13.

Added to machdep.c system clock configuration and cpu identification.

Add stm32f411xe.h device header from STM32CubeF4 V1.24.2 (Dec 06, 2019).
The header file has been run through dos2unix and whitespace removal.

Added the NUCLEO-F411RE development board BOARD define to the root's
Makefile.inc for running OpenOCD and GDB, such as:
  $ BOARD=f411renucleo gmake ocd
  $ BOARD=f411renucleo gmake gdb-ocd
A connection through the VCP serial port, on OpenBSD, via cu:
  $ cu -l /dev/cuaU0 -s 115200