sys/{pic32,stm32}: Use sd controller driver name from *config.
The sd driver uses an spi controller on pic32, or it uses an spi
or an sdio controller on stm32.
Obtain the name of the sd controller driver ("spi" or "sdio")
directly through the *config struct parameter in sd_probe().

Both ports handle the name in a unified fashion.

Remove "4-bit bus" from stm32 sd_probe() for now, as that only
applies to an sdio controller and not to an spi controller.