src/libtcl: Revive libtcl 6.7; some unsigned char * fixes.
The tcl 6.7 library was never enabled in the RetroBSD source tree.
Hook it into the DiscoBSD build now, for both pic32 and stm32.

A build with tclsh gives an executable size of about 67kbytes on
stm32, and a size of about 73kbytes on pic32 (shell is 70kbytes).

Some unsigned char * to char * fixes, so the library builds cleanly.
The whole library will eventually need a fixup of the unsigned char *
issues present in the API, or rather be replaced by a newer version.

Note that this is only availing libtcl to cross compiled tcl programs
(namely tclsh) built and installed into the DiscoBSD root file system.
A separate commit will enable the tcl library proper to be built and
included in the system with the other libraries.