cmd/tclsh: Enable tclsh for all ports; new Makefile.
For both pic32 and stm32 ports, tclsh built with libtcl 6.7 runs well.
tclsh is a very minimalist version of the tcl interactive shell.

Keeping -Wno-pointer-sign CFLAG until libtcl unsigned/signed char *
issues are fixed.

Added a .gitignore for the tclsh binary.

tclsh is added to the base machine-independent manifest set to be
included in the root file system under /bin/tclsh, which is in PATH.