README.md: Add text for history, SD card imaging, pic32 dmesg.
Add a History section to explain how DiscoBSD came about and the
current status of the operating system.
Add links to RetroBSD and the directed study and the paper.

Add more detailed instructions on imaging an SD card with sdcard.img.

PIC32 kernels are normally flashed via a hex file, not an elf file.

Add links to the stm32 and pic32 README.md files.

Update the DiscoBSD/stm32 dmesg to the kernel in the 2.0 release.
Add the DiscoBSD/pic32 dmesg of the kernel in the 2.0 release.