src/libreadline: readline subdir for lib header files.
As the GNU Readline library installs its header files at the location of
.../lib/readline/{readline,history}.h, do the same for the Linenoise-based
Readline-like library. Adjust all uses of the readline library in the tree.

pdc now explicitly pulls in readline through the source tree via
CPPFLAGS += -I../../libreadline

The files Makefile-unix and example.c are updated to the header file's
new directory, but are not used nor tested.