build: etc/Makefile distrib-dirs creates DESTDIR tree.
Bring back use of etc/Makefile and add the new target 'distrib-dirs'
that builds an installation directory tree at DESTDIR, as below:
|-- bin
|-- etc
|-- games
|   `-- lib
|-- include
|-- lib
|-- libexec
|-- root
|-- sbin
|-- share
|   `-- man
|       |-- cat1
|       |-- cat2
|       |-- cat3
|       |-- cat4
|       |-- cat5
|       |-- cat6
|       |-- cat7
|       `-- cat8
`-- var
    `-- log

The end goal of this work is to install into DESTDIR where DESTDIR
is not co-located with the source code. This enables multi-platform
builds depending on ${MACHINE} or another location for ${DESTDIR},
such as DESTDIR ?= ${TOPSRC}/distrib/obj/destdir.${MACHINE}

Added DESTDIR ?= ${TOPSRC} to the source tree's Makefile, so as not
to pull in target.mk. Both DESTDIR defines shall be kept in sync.

Currently ${DESTDIR} == ${TOPSRC}, so "mkdir -p" is a no-op.
Directories should be created owner root group wheel, but that
requires build users to have elevated permissions.
DiscoBSD is cross-compiled, so fsutil will create the root file system
with proper owners, groups, and permissions for files and directories.

Inspiration from OpenBSD, NetBSD, and 2.11BSD.