src/share/zoneinfo: No local tzfile.h from ${DESTDIR}.
The file ${TOPSRC}/include/tzfile.h, required for zic and zdump,
can be included via a -idirafter include directory added to CFLAGS.

It is not needed anymore to firstly install tzfile.h in ${DESTDIR}
and then soft link to it before compilation.

This enables a full src tree compile of src objects, without first
needing to install anything.

This returns/reverts the zoneinfo Makefile to something closer to
what was shipped in 2.11BSD.

Remove tzfile.h file from .gitignore.