build: Move /lib directory to src tree at /src/libc_aout.
Bring the /lib directory subtree into the source tree as /src/libc_aout.
The libraries under /src/libc_aout are for installation into
${DESTDIR}/lib and used with the installed a.out-based development tools.

Turn off making, installing, and cleaning the /lib source directory.

This commit has no modifications to the library Makefiles. That will
come in a future commit.

Note that a "gmake fs" will output errors for missing libraries in /lib,
which is to be expected with this commit and remedied in the next one.