etc/motd: Install etc.${MACHINE}/motd with etc/Makefile.
No need to specially handle motd anymore. Install it through the
'distribution' target in etc/Makefile as one of the ${FILES},
dependent on ${MACHINE}, which is passed by the gmake -C etc recipe.

Remove motd from .gitignore, along with the rest of etc/.gitignore
and readline from include/.gitignore, that are no longer installed
in the source tree during a build.