build: 4.4BSD hierarchy, move /src/share to /share.
Further towards the goal of a source tree in the 4.4BSD hierarchy,
and since ${DESTDIR} installs are fully enabled,
start the process of moving directories into their preferred place.

The root Makefile builds and installs for /share.
Adjust the relative location of ${TOPSRC} in Makefiles.

All other files are renamed, without changes to their file contents.