build: 4.4BSD hierarchy, move /src/cmd bins to /bin.
The /bin source directory is for binaries needed when running
in single-user mode. The list of binaries is similar to that
of Net/2, 4.4BSD, OpenBSD, and NetBSD.
Binaries are removed from /src/cmd/Makefile and added to the
new /bin/Makefile (based on the /src/cmd and 2.11BSD Makefile).

Enable /bin build, install, and clean via the root Makefile.

Adjust /src/cmd/.gitignore and add a /bin/.gitignore file.

Adjust the relative location of ${TOPSRC} in the /bin Makefiles.
All other files are renamed, without changes to their contents.