build: 4.4BSD hierarchy, move /src/cmd sbins to /sbin.
The /sbin source directory is for binaries for the root user
that are needed when running in single-user mode.

The list of binaries is similar to that of Net/2, 4.3BSD-Reno,
4.4BSD, OpenBSD, and NetBSD.

For example, mkpasswd is in /sbin in 2.11BSD but is in /usr/sbin
in 4.3BSD-Reno, so it goes in /usr/sbin in DiscoBSD, not /sbin.

Binaries are removed from /src/cmd/Makefile and added to the
new /sbin/Makefile (based on the deleted Makefile-sbin and
the original 2.11BSD /sbin/Makefile).

Enable /sbin build, install, and clean via the root Makefile.

Adjust the relative location of ${TOPSRC} in the /sbin Makefiles.
All other files are renamed, without changes to their contents.