build: 4.4BSD hierarchy, move /src/cmd sbins to /usr.sbin.
The /usr.sbin source directory is for binaries for the root user
in multi-user mode, as /usr may be mounted independently of the
root file system and is not mounted in single-user mode.

The list of binaries is similar to that of Net/2, 4.3BSD-Reno,
4.4BSD, OpenBSD, and NetBSD.

devupdate is included in /usr.sbin but is not yet in the build.

Binaries are removed from /src/cmd/Makefile and added to the
new /usr/sbin/Makefile (based on /src/cmd and 2.11BSD Makefiles).

Enable /usr.sbin build, install, and clean via the root Makefile.

Adjust /src/cmd/.gitignore and add a /usr.sbin/.gitignore file.

Adjust the relative location of ${TOPSRC} in /usr.sbin Makefiles.
All other files are renamed, without changes to their contents.