build: 4.4BSD hierarchy, rename /src to /lib.
The /lib source directory is for system libraries.

Enable /lib build, install, and clean via the root Makefile.

Paths in Makefiles that specified 'src' are converted to 'lib'.
Relative paths need no change.
All other files are renamed, without changes to their contents.

This completes the major restructuring of the DiscoBSD source
tree root-level directories to be in line with 4.4BSD and its
derivatives, such as OpenBSD and NetBSD.

The DiscoBSD source tree root directories are as follows:
  ${TOPSRC}
  |-- bin
  |-- distrib
  |-- etc
  |-- games
  |-- home
  |-- include
  |-- lib
  |-- libexec
  |-- sbin
  |-- share
  |-- sys
  |-- tools
  |-- usr.bin
  `-- usr.sbin