usr: Bring /libexec up to /usr/libexec in file system.
This is the start of bringing certain file system
directories back into /usr where they belong.

The first directory hoisted up to /usr is /libexec, up to
/usr/libexec, with all of its executables, such as getty.

Directories in the fsutil file system manifests are updated
with the /usr prefix on all the /libexec lines.

Create /usr/libexec in the /etc/Makefile distrib-dirs target.

/etc/ttys adjusted location for getty as /usr/libexec/getty.

Makefile install targets adjusted with /usr prefix for libexec.

Compilers and whereis adjusted with /usr prefix for libexec.