usr: Bring /lib up to /usr/lib in file system.
Since the /bin and /sbin utilities are statically compiled, there
is no need for system libraries to be in /lib in single-user mode.
This follows precedent set by 4.4BSD, OpenBSD, and others.

Hoist /lib up to /usr/lib, and adjust all paths accordingly.

Directories in the fsutil file system manifests are updated
with the /usr prefix on all the /lib lines.
The /lib directory is no longer created in the file system.

Create /usr/lib in the /etc/Makefile distrib-dirs target.

Makefile install targets adjusted with /usr prefix for libraries.

Source files and Makefiles are not updated for utilities that are
not yet in the build, such as ccom, lccom, pforth, and unixbench.
Manual pages are not updated either, but many are correct, as
they were originally written with libraries being in /usr/lib.

ld updated to search for libraries in /usr/lib instead of /lib.