usr: Bring /games up to /usr/games in file system.
Hoist /games up to /usr/games, /games/lib up to /usr/games/lib,
and adjust all paths accordingly in Makefiles and source files.

Directories in the fsutil file system manifests are updated
with the /usr prefix on all the /games and /games/lib lines.
The /games directory is no longer created in the file system.

Create /usr/games and /usr/games/lib in the /etc/Makefile
distrib-dirs target.

Makefile install targets adjusted with /usr prefix for games.

Some manual pages updated while in there, but many are correct
as it is, as this commit is restoring the 2.11BSD structure.

Having lib at /usr/games/lib is in line with 4.3BSD-Tahoe and
2.11BSD. Future commits will follow 4.3BSD-Reno, 4.4BSD, and
others where those files are in /usr/share/games or /var/games.