usr: Bring /{,s}bin up to /usr/{,s}bin in file system.
Hoist multi-user mode programs in /bin and /sbin up to /usr/bin
and /usr/sbin, respectively, following 4.4BSD, OpenBSD, and others.
Adjust all paths accordingly in Makefiles and source files,
especially for executables such as /usr/bin/login.
Adjust the paths of PATH in /etc/rc, etc/passwd, etc/shadow,
dot.profile, and include/paths.h.

Directories in the fsutil file system manifests are updated
with the /usr prefix on the /bin and /sbin lines. Care has been
given to not rearrange the lines, to give clarity to the diff.
A future commit will rearrange the manifests based on directory.

Of note, iostat and sysctl have moved from /bin to /usr/sbin,
updatedb has moved from /sbin to /usr/libexec, and all of the uucp
suite has been moved (for now) from /bin to /usr/bin, eventhough
some should be in /usr/sbin or /usr/libexec, as the case may be.

Most programs that are not in the build have been ignored.
Manual pages will be updated in a separate commit.

This leaves the file system in quite a comfortable 4.4BSD state.