sys/stm32: Portable sys/stm32/gcc-config.mk makefile.
A few tricks to remove the gmake-isms from gcc-config.mk and make
it portable with gmake and the various BSD makes.

Determine the host operating system name to select the path to the
GCC arm-none-eabi compiler tools, and fail with /does/not/exist.

!= runs the shell commands and saves the output to ARM_GCC_PREFIX.

On the command line a different tool prefix can be used like:
  $ make ARM_GCC_PREFIX="/path/to/my/arm-none-eabi"

Instructions on installing the toolchain on OpenBSD are available
from tools/openbsd/README.md in the repository.