sys/pic32: Portable GCC sys/pic32/gcc-config.mk makefile.
Remove all old and untested toolchains and keep only GCC.
Toolchains can be re-added once they are vetted.

A few tricks to remove the gmake-isms from gcc-config.mk and make
it portable with gmake and the various BSD makes.

Determine the host operating system name to select the path to the
GCC mips-elf compiler tools, and fail with /does/not/exist-.

!= runs the shell commands and saves the output to MIPS_GCC_PREFIX.

On the command line a different tool prefix can be used like:
  $ make MIPS_GCC_PREFIX="/path/to/my/mips-elf-"