sys/pic32: Portably compile devcfg.c if present.
Use the shell to determine if devcfg.c exists and include devcfg.o
in ${SYSTEM_OBJ}, or else include just an empty string.

This removes the ifeq() gmake-ism and allows the makefile to be
portable with gmake and the various BSD makes.

Remove references to MIPS_GCC_FORMAT, as the linker scripts use
OUTPUT_FORMAT(), which makes MIPS_GCC_FORMAT redundant.

Regen makefiles in next commit.