build: Kernel build makefiles are fully portable.
The kconfig kernel system spec is now 'unix', not 'unix.elf'.

Fix the fallout of removing the .elf suffix in the various README files.

Remove the ifeq() for dependency files, as the -include fails gracefully
if the files do not exist, so no need to check.

stm32 kernels do not need the pic32-specific 'load' target.

No need to use the $(basename $@) gmake-ism; just use plain $@.

.gitignore 'unix' and regen makefiles in next commit.