build: Portable GCC-only target.mk makefile.
Remove all old and untested MIPS toolchains and keep only GCC.
Toolchains can be re-added once they are vetted.

A few tricks to remove gmake-isms from target.mk and make
it portable with GNU make and the various BSD makes.

Replace the $(shell ...) gmake-ism with variables derived from
running the shell with the "!=" variable assignment operator.

The structure and commands have largely stayed the same as before,
as this commit is mainly mechanical in nature.

The "-" at the end of ${GCCPREFIX} is no longer; it is now the
suffix's responsibility. Adjust all uses of ${GCCPREFIX} accordingly.