build: Handle ${YACC} and ${YFLAGS} variables consistently.
Default value of the build variable YFLAGS is '-d' in target.mk
to cause the header file 'y.tab.h' to be written.

YFLAGS is set to nothing for those instances where the 'y.tab.h'
header file is not needed.

Set YACC variable as only the executable run for calls to yacc,
and do not include any flags in the YACC variable.
Flags all go in the YFLAGS variable.

Consistently use 'rm y.tab.[ch]' in make clean targets, whether
or not y.tab.h is generated or y.tab.c has been renamed.

New .gitignore file for lex to ignore 'y.tab.c'.

The only altered behaviour in the commit is that the '-d' flag
does not need to be used for building pdc, lccom, and make.