build: Portable makefile non-suffix rule contexts.
Replace the GNU make idiom of using "$<" in a non-suffix rule
context with the actual name of the prerequisite source file.

This change makes these Makefiles portable between GNU make
and the various BSD makes.

Replace all calls to 'nroff' to use the ${MANROFF} variable.

unixbench and lccom are absent from this commit, as they are
not yet even hooked into the build.
The messier parts of libc_aout will be handled separately.