usr.bin/picoc: Portable Makefile build to current directory.
Portably build picoc with objects in the current directory with
both GNU make and BSD make.

Do not build objects in source subdirectories, as previously done,
by forcing OBJS and dependencies to be the current directory.

Find sources in subdirectories through VPATH= (for GNU make) and
.PATH: (for BSD make). A clever trick which works for both makes.

Spell out objects in ${OBJS} to be in the current directory.

The whole source tree, excepting libc_aout, can now be built with
both GNU make and BSD make.