lib/libc: Use toolchain ${AR}, not host's default ar.
Be consistent with only using the toolchain ${AR} when building
libraries. Note that this does not apply to building lib/libc_aout
libraries, which use the in-tree usr.bin/ar running on the host.

Remove the leftover DESTDIR variables in libtermlib/termcap and
net/named libraries. DESTDIR is to be defined in /target.mk and
/Makefile only.