build: 4.4BSD hierarchy, own directory for whereis.
4.4BSD (due to pmake) requires executables to live in their own
source directories, while 4.3BSD-Tahoe allowed single-source-file
executables to live in the common subdirectory, such as /usr.bin.

The 4.4BSD approach enables read-only sources and builds into a
softlinked ./obj directory. This is a step towards that reality.

Follow the 4.4BSD requirement of independent directories, and
start with /usr.bin/whereis as the first example.
This Makefile will be a template for the executables that follow.
Executables need only modify the Makefile ${PROG} variable name.

Add a .gitignore file for ignoring the compiled executable.