build: zoneinfo /etc/localtime handled by /etc/Makefile.
The zoneinfo file installed at /etc/localtime is now the
responsibility of /etc/Makefile, which handles the 'distribution'
make target that requires super user privileges for installation.

Forcefully make a soft link of the default zoneinfo time zone.

TZDIR and LOCALTIME in /etc/Makefile and /share/zoneinfo/Makefile
need to be kept in sync. They are not used anymore in
/share/zoneinfo/Makefile, but are kept for completeness sake.