sys/stm32: Kernel build with -fcommon, without -nopie.
Build the stm32 kernels with -fcommon added to CMACHFLAGS and
-nopie removed from LDFLAGS in Makefile.kconf.

Regenerate stm32 kernel Makefiles.

-nopie (and/or -no-pie or the many other variations) is the
default for arm-none-eabi-{gcc,ld}, so removing it does not
change the produced kernel binary.

Prior to GCC 10 -fcommon was the default. GCC 10+ defaults to
-fno-common. Use -fcommon until the kernel is -fno-common clean.
See https://gcc.gnu.org/git/gitweb.cgi?p=gcc.git;h=6271dd984d7f92

This commit does not change the produced kernel binary with:
- OpenBSD/amd64 6.8 and 7.3, with arm-none-eabi-gcc-linaro 7.4.1.
- Ubuntu/amd64 18.04 LTS, with gcc-arm-none-eabi 7.3.1.

This commit fixes "-nopie not an option" errors and "multiple
definition of" errors with:
- FreeBSD/amd64 13.2, with both arm-none-eabi-gcc 11.3.0 and
  gcc-arm-embedded 10.3.1.
- Ubuntu/amd64 23.04, with gcc-arm-none-eabi 12.2.1.

Modern FreeBSD and Ubuntu are now able to build stm32 kernels.

This commit is inspired/based on previous work by @lhondareyte in
pull request #2 and @saper in issue #4, and the -fno-common work
done in OpenBSD for base clang/llvm.