tools: Implement ${HOST_CC} make variable; default cc.
The ${HOST_CC} make variable is the compiler name on the host
operating system to build the /tools, and can be changed via:

  $ make HOST_CC=gcc

when building the system. The default compiler is cc. Base system
compiler names vary, but cc is the de facto standard name in Unix.

Start using the 4.4BSD style ../Makefile.inc includes, so common
constructs can be unified and consistent in one common file.