build: Implement ${HOST_CC} make variable; default cc.
Implement the ${HOST_CC} make variable for the rest of the build.
It is the name of the compiler used on the host operating system.
A different compiler than the default, cc, can be used via:

  $ make HOST_CC=egcc

when building the system.

Programs not yet hooked into the build, such as lccom, lex, and m4,
are not updated in this commit to use ${HOST_CC}.

Note that lib/libc_aout/Makefile does not source /target.mk, and
so needs to define the ${HOST_CC} variable on its own.
${CC} is defined as ${HOST_CC} so that default compile rules will
use the specified host compiler in building the aout utilities.