sys/stm32: Use devel/gcc-arm-embedded on FreeBSD.
The devel/arm-none-eabi-gcc package on FreeBSD does not include the
required multilib Cortex-R/M profile, and so generates executables with
instructions that are unable to run on an STM32F4 device (or any
ARMv7E-M device for that matter).

Build stm32 kernels with the FreeBSD devel/gcc-arm-embedded package.

Refer to Pull Request #10 for more details.