tools/aoututils: Bring a.out utilities into /tools.
New build structure in /tools for a.out utilities, to build
the in-system /usr/lib libraries installed in DiscoBSD.

The a.out utilities are: aout ar as ld nm ranlib size strip.

A Makefile.inc file provides common defines and productions.
Need only to provide ${PROG} and ${OBJS} in local Makefile.

The a.out utilities are installed into ${TOOLBINDIR}.

The source code for each utility has been copied from usr.bin,
as to keep the build system cleaner and easier to manage.
The source code in both places must be kept synchronized.
A NetBSD-style "reachover" mechanism may be used in the future.

A local "include" directory is used for the a.out header files.

New .gitignore file for each utility.

These utilities are not yet utilized in the build - that's for
another commit.