libc: ANSIfy vararg function definitions in {gen,stdio}.
GCC's -Wattributes triggers on 'format' attributes on variable
argument functions in K&R C, and thinks they are not varargs.

Converting to ANSI function definitions clears it all up.

Note the small change in private function _conv() where the 'pad'
argument was an int in K&R, but is now a char in keeping with the
4.3BSD-Reno version of strftime.c.

Functions are: strftime() and v{,f,s}printf().