usr.bin/smlrc: Do not check string against NULL in strlen().
The strlen() function must be supplied with a proper string,
and NULL is not a string. If NULL is supplied to strlen() then
the result is undefined behaviour, and returning zero (as this
library would do) is completely incorrect. A proper string with
a length of zero is just the terminating NUL byte, such as "".

Any program that fails due to this change was already broken in
the first place, and this function only covered up the error.

The correct behaviour has been validated by the implementations
of strlen() in DiscoBSD libc and kernel, libc and libkern for
OpenBSD and NetBSD, and P.J. Plauger's The Standard C Library.

The C Standard specifies in 7.1.4.1: "If an argument to a
function has an invalid value ... the behaviour is undefined."

Clears -Wnonnull-compare warning from mips-elf-gcc 12.2.0 build.