build: Enable -fcommon for base and kernel builds for mips.
gcc 10+ defaults to -fno-common, which produces many "multiple
definition of ..." errors throughout the tree.

Enable -fcommon for mips/pic32 until tree is -fno-common clean.

Modern OpenBSD, FreeBSD, and Ubuntu are now able to build the
base and kernel for pic32 with a mips-elf-gcc 10+ compiler.

Remove the dubious -fshort-double CFLAG, as it was removed from
mips-elf-gcc a while ago and modern compilers do not support it.