usr.bin/lex: Import MIT-licensed Unix lex from Plan 9.
Import the original Unix lex from Plan 9 from Bell Labs,
Fourth Edition from August 03, 2011, verbatim.

Plan 9 has been licensed as MIT by the Plan 9 Foundation since 2021.

This version of lex is largely the same as the original AT&T lex
in 2.11BSD, with modifications to run on Plan 9. These changes
will be reverted (mainly file handling and printing), so that
this lex will run on DiscoBSD.

This does not yet build or run, and is not hooked into the build.

Major changes of this lex versus the lex shipped in 2.11BSD is:
ansi functions, more precise types and typecasting, removal of
support for Ratfor, 8-bit characters, use C library functions,
and yylval is a union.