usr.bin/lex: Move lex scanner to /usr/share/misc/ncform.
Further to the goal of normalizing the file system, and
/usr/share in particular, move the lex scanner skeleton
to /usr/share/misc/ncform where it is beside the yacc parser.

Update the new location in the fsutil spec for md.stm32, the
cname variable in lex, and the lex Makefile install target.

The chosen location is the same as in 4.3BSD-Reno and 4.4BSD,
but was at /usr/libdata/lex/ncform in 2.11BSD and was at
/usr/lib/lex/ncform in 4.3BSD-Tahoe.