usr.bin/re: Move re.help to /usr/share/misc/re.help.
Further to the goal of normalizing the file system, and
/usr/share in particular, move the key bindings file for
re to /usr/share/misc/re.help.

This change is also a bugfix. The re editor will open and
display the file re.help when it starts, but the old path
for deffile[] was no longer correct, since it was missed
during the /usr hoisting. It will now display correctly.

Update the new location in the fsutil spec for base/mi, the
deffile[] variable in re, and the re Makefile install target.

/usr/share in the file system is now a directory of only
directories. All files are located in subdirectories.