usr.bin/re: Off by one bugfix on status line overflow.
The terminal width is derived from NCOLS (normally 80 wide).
The status bar has a fixed-size status on the lower right
side of the terminal that is defined by PARAMWIDTH and is
actually 19, not 18, characters wide.

The text printed is:
~~Ins~~Line:~~~~~~~
or
~~~~~~~Line:~~~~~~~
which clearly is 7 + 5 + 7 = 19.

The status bar now displays properly, without overflowing.