usr.bin/bc: Import MIT-licensed Unix bc from Plan 9.
Import the original Unix bc from Plan 9 from Bell Labs,
Fourth Edition from September 08, 2009, verbatim.
Import the manual page as well.

Plan 9 has been licensed as MIT by the Plan 9 Foundation since 2021.

This version of bc is largely the same as the original AT&T bc
in 2.11BSD, with modifications to run on Plan 9. These changes
will be reverted in a separate commit (mainly file handling and
printing), so that this bc will run on DiscoBSD.

This does not yet build or run, and is not hooked into the build.
Temporarily remove /usr/bin/bc, /usr/share/misc/lib.b, and
/usr/share/man/cat1/bc.0 from the fsutil spec for base/mi.

Major changes of this bc versus the bc shipped in 2.11BSD is:
cleanup of reduce/reduce errors in the grammar, removal of the
old style of using "=" in the productions, add -s option to
suppress automatic display of calculation results, add -d option
for debugging output, bugfix for lib.b ibase save and restore,
yylval is a union, and general code style clean up and fixes.