usr.bin/bc: Convert to Unix print and file i/o functions.
Replace Plan 9-specific functions with Unix-specific ones,
such as Bopen() to fopen() and Bprint() to printf().

Use the DiscoBSD file system locations for lib.b and dc.

bc is now compilable and usable; although it is not yet
hooked into the build, it will be in a separate commit.

The Plan 9 versions of bc and dc are uniquely entwined,
so the Plan 9 dc utility may also need to be ported.