build: Enable /usr/bin/bc; generate manual page.
Enable build and installation of /usr/bin/bc for all architectures.
Generate the bc manual page local to the program, and install it.

Note that this version of bc (from Plan 9) requires the matching
version of Plan 9 dc that understands crs in angle brackets, so
until dc is ported this bc may be of limited use and functionality.