usr.bin/dc: Import MIT-licensed Unix dc from Plan 9.
Import the original Unix dc from Plan 9 from Bell Labs,
Fourth Edition from May 03, 2011, verbatim.
Import the manual page as well.

Plan 9 has been licensed as MIT by the Plan 9 Foundation since 2021.

This version of dc is largely the same as the original AT&T dc
in 2.11BSD, with modifications to run on Plan 9. These changes
will be reverted in a separate commit (mainly file handling and
printing), so that this dc will run on DiscoBSD.

This does not yet build or run, and is not hooked into the build.
Temporarily remove /usr/bin/dc, and /usr/share/man/cat1/dc.0 from
the fsutil spec for base/mi.

Major changes of this dc versus the dc shipped in 2.11BSD are:
able to accept the output of Plan 9 bc which has crs in angle
brackets, update K&R syntax to ANSI, formatting cleanup, bug
fixes, and many grammar fixes in the manual page.