usr.bin/dc: Convert to Unix functions; enable /usr/bin/dc.
Replace Plan 9-specific functions with Unix-specific ones,
such as Bopen() to fopen() and Bprint() to printf().

Use /bin/sh for shell commands instead of Plan 9 /bin/rc/.

Fix bug that would cause crashes due to DiscoBSD's strict
malloc() implementation. A call to salloc() with a size
of zero now increases the size to at least one, so the
minimum size request is malloc(1), not malloc(0).
This is also in line with the old behaviour of dc.

Enable the building and installation of /usr/bin/dc and
its manual page for all architectures.

Clean up and normalize Makefile.

This Plan 9 version of dc is a derivative of the original
Unix dc and works with the ported version of Plan 9 bc,
but may or may not work well with other implementations
such as those from the FSF or OpenBSD.

bc can now take full advantage of using dc for calculations.

bc and dc pass many of the OpenBSD regress suite tests.