usr.bin/whereis: Import whereis manual page from 4.2BSD.
Import, generate, and install the whereis.1 manual page from
4.2BSD, as it seems to be the most "full" version available.

This manual page is specifically for whereis(1) in 2.11BSD,
written by Bill Joy, and not any of the later re-writes.

Update Makefile to generate and install the manual page.

Add /usr/share/man/cat1/whereis.0 to the base/mi fsutil spec.