usr.bin/whereis: Use relevant example in manual page.
Clean up the order of options into alphabetical order.
Update the usage() in whereis.c as well.

Use an example that is relevant for the DiscoBSD file system
hierarchy, and is tested to work.

Clean up the old 2.11BSD and 4.3BSD-Tahoe directories, as
they are not used here.

Add proper date string to manual page.