build: Remove unused and unsupported cross.mk.
The "cross" building structure inherited from RetroBSD
for building host executables is not used, Linux specific,
and potentially dangerous to the host environment when
certain options are set or are not set. Be rid of it.

Rid the source tree of all vestiges of cross.mk, mainly
as a commented-out included file in Makefiles.

Clean up some CFLAGS with -DCROSS while at it.

If an executable needs to be compiled to run on the host
machine, then use the structure provided in /tools.