build: Add back -fshort-double build option for mips.
The modern mips-elf-gcc compilers don't handle short doubles,
which the pic32 mips uses, and the quad floating support or
runtime support has not yet been added to handle this case.

The assembler in tools/aoututils, to build the installed libraries,
does not yet handle .module from the newer gcc versions.

Keep using the old 4.8.1 gcc compiler for now with the 2.27
version of binutils. The binutils 2.39 version also seems okay.