build: Move target.mk to share/mk/sys.mk; update Makefiles.
Makefiles are updated with the new file location for the
system Makefile, located at ${TOPSRC}/share/mk/sys.mk and
included in almost all Makefiles throughout the source tree.

The file location was moved to be more in line with historical
and contemporary 4.4BSD usage, and to be ready for a potential
switch to pmake (the BSD parallel make).

No change in sys.mk file contents. Only a rename.

Currently sys.mk is not the same as the POSIX sys.mk files found
in the modern BSDs, but it will be soon. This is the first step.