usr.bin/retroforth: Shorten filenames in rosetta_code.
The names of files in the rosetta_code examples are based on the
naming of the same examples from the Rosetta Code website.

Two of the names on the website have been given new names
(...dimensions.rx, and ...cursor_positioning.rx) while the other
(...ring_the_bell.rx) just needed to be shortened.

There is a max filename length for the old style of tar archives,
which is currently used in DiscoBSD. These long filenames are
skipped by DiscoBSD's tar when extracting the source tree archive.

They have been shortened, and now are properly handled with tar.